BO_Package/requirements.txt
# Bayesian Optimizer Package - Python Dependencies

# Bayesian Optimization
optuna>=3.0.0

# Kafka Integration
kafka-python>=2.0.2

# HTTP Client
requests>=2.28.0

# Numerical Computing
numpy>=1.21.0
pandas>=1.3.0

# Data Validation (optional, for future enhancements)
# pydantic>=1.10.0
